Cleaned up big time!!
There was 2 api folders for some reason one in ThreadUD/api/ one in ThreadUD/backend/api which we do not need the 2 of them so I cleaned that up also in the backend/config/ I made a file called apiConfig.js that handles the api url so that we only have to change it in once place when someone else works on the project. I cleaned up the index.tsx from the mess I left last time and I also added separate style sheets for cleanness so it's not all in the main page's file cuz that can get messy later on.